🐛 Fix intro process
