+++ lib/ui/views/intro/layouts/intro_configure_security.dart
@@ -1,33 +1,21 @@
-import 'dart:async';
-import 'dart:developer';
-
 import 'package:aewallet/application/authentication/authentication.dart';
 import 'package:aewallet/application/connectivity_status.dart';
 import 'package:aewallet/application/settings/settings.dart';
-import 'package:aewallet/application/wallet/wallet.dart';
 import 'package:aewallet/bus/authenticated_event.dart';
-import 'package:aewallet/bus/transaction_send_event.dart';
-import 'package:aewallet/infrastructure/datasources/hive_vault.dart';
 import 'package:aewallet/model/authentication_method.dart';
-import 'package:aewallet/model/data/appdb.dart';
 import 'package:aewallet/ui/themes/archethic_theme.dart';
 import 'package:aewallet/ui/themes/styles.dart';
-import 'package:aewallet/ui/util/ui_util.dart';
 import 'package:aewallet/ui/views/authenticate/pin_screen.dart';
 import 'package:aewallet/ui/views/intro/bloc/provider.dart';
-import 'package:aewallet/ui/views/main/home_page.dart';
 import 'package:aewallet/ui/views/settings/set_password.dart';
 import 'package:aewallet/ui/views/settings/set_yubikey.dart';
 import 'package:aewallet/ui/widgets/components/icon_network_warning.dart';
 import 'package:aewallet/ui/widgets/components/picker_item.dart';
 import 'package:aewallet/ui/widgets/components/scrollbar.dart';
-import 'package:aewallet/ui/widgets/components/show_sending_animation.dart';
 import 'package:aewallet/ui/widgets/dialogs/authentification_method_dialog_help.dart';
 import 'package:aewallet/util/biometrics_util.dart';
 import 'package:aewallet/util/get_it_instance.dart';
 import 'package:aewallet/util/haptic_util.dart';
-import 'package:aewallet/util/keychain_util.dart';
-import 'package:archethic_lib_dart/archethic_lib_dart.dart';
 import 'package:auto_size_text/auto_size_text.dart';
 import 'package:event_taxi/event_taxi.dart';
 import 'package:flutter/material.dart';
@@ -58,146 +46,8 @@ class _IntroConfigureSecurityState
   PickerItem? _accessModesSelected;
   bool? animationOpen;
 
-  StreamSubscription<TransactionSendEvent>? _sendTxSub;
-  StreamSubscription<AuthenticatedEvent>? _authSub;
-  bool keychainAccessRequested = false;
-  bool newWalletRequested = false;
-
-  void _registerBus() {
-    _authSub = EventTaxiImpl.singleton()
-        .registerTo<AuthenticatedEvent>()
-        .listen((AuthenticatedEvent event) async {
-      await createKeychain();
-    });
-
-    _sendTxSub = EventTaxiImpl.singleton()
-        .registerTo<TransactionSendEvent>()
-        .listen((TransactionSendEvent event) async {
-      final localizations = AppLocalizations.of(context)!;
-
-      if (event.response != 'ok' && event.nbConfirmations == 0) {
-        UIUtil.showSnackbar(
-          '${localizations.sendError} (${event.response!})',
-          context,
-          ref,
-          ArchethicTheme.text,
-          ArchethicTheme.snackBarShadow,
-        );
-
-        return;
-      }
-
-      if (event.response != 'ok') {
-        UIUtil.showSnackbar(
-          localizations.notEnoughConfirmations,
-          context,
-          ref,
-          ArchethicTheme.text,
-          ArchethicTheme.snackBarShadow,
-        );
-        return;
-      }
-
-      switch (event.transactionType!) {
-        case TransactionSendEventType.keychain:
-          UIUtil.showSnackbar(
-            event.nbConfirmations == 1
-                ? localizations.keychainCreationTransactionConfirmed1
-                    .replaceAll('%1', event.nbConfirmations.toString())
-                    .replaceAll('%2', event.maxConfirmations.toString())
-                : localizations.keychainCreationTransactionConfirmed
-                    .replaceAll('%1', event.nbConfirmations.toString())
-                    .replaceAll('%2', event.maxConfirmations.toString()),
-            context,
-            ref,
-            ArchethicTheme.text,
-            ArchethicTheme.snackBarShadow,
-            duration: const Duration(milliseconds: 5000),
-            icon: Symbols.info,
-          );
-
-          if (keychainAccessRequested) break;
-
-          setState(() {
-            keychainAccessRequested = true;
-          });
-          await KeychainUtil().createKeyChainAccess(
-            ref.read(SettingsProviders.settings).network,
-            widget.seed,
-            event.params!['keychainAddress']! as String,
-            event.params!['originPrivateKey']! as String,
-            event.params!['keychain']! as Keychain,
-          );
-          break;
-        case TransactionSendEventType.keychainAccess:
-          UIUtil.showSnackbar(
-            event.nbConfirmations == 1
-                ? localizations.keychainAccessCreationTransactionConfirmed1
-                    .replaceAll('%1', event.nbConfirmations.toString())
-                    .replaceAll('%2', event.maxConfirmations.toString())
-                : localizations.keychainAccessCreationTransactionConfirmed
-                    .replaceAll('%1', event.nbConfirmations.toString())
-                    .replaceAll('%2', event.maxConfirmations.toString()),
-            context,
-            ref,
-            ArchethicTheme.text,
-            ArchethicTheme.snackBarShadow,
-            duration: const Duration(milliseconds: 5000),
-            icon: Symbols.info,
-          );
-
-          if (newWalletRequested) break;
-
-          setState(() {
-            newWalletRequested = true;
-          });
-          var error = false;
-          try {
-            await ref
-                .read(SessionProviders.session.notifier)
-                .createNewAppWallet(
-                  seed: widget.seed!,
-                  keychainAddress: event.params!['keychainAddress']! as String,
-                  keychain: event.params!['keychain']! as Keychain,
-                  name: widget.name,
-                );
-          } catch (e) {
-            error = true;
-            UIUtil.showSnackbar(
-              '${localizations.sendError} ($e)',
-              context,
-              ref,
-              ArchethicTheme.text,
-              ArchethicTheme.snackBarShadow,
-            );
-          }
-          if (error == false) {
-            context.go(HomePage.routerPage);
-          }
-          break;
-        case TransactionSendEventType.transfer:
-          break;
-        case TransactionSendEventType.token:
-          break;
-      }
-    });
-  }
-
-  void _destroyBus() {
-    _sendTxSub?.cancel();
-    _authSub?.cancel();
-  }
-
-  @override
-  void dispose() {
-    _destroyBus();
-    super.dispose();
-  }
-
   @override
   void initState() {
-    log('initstate');
-    _registerBus();
     animationOpen = false;
     super.initState();
   }
@@ -418,40 +268,6 @@ class _IntroConfigureSecurityState
         ),
       ),
     );
-  }
-
-  Future<void> createKeychain() async {
-    final localizations = AppLocalizations.of(context)!;
-
-    ShowSendingAnimation.build(
-      context,
-      title: localizations.appWalletInitInProgress,
-    );
-
-    try {
-      await sl.get<DBHelper>().clearAppWallet();
-      final vault = await HiveVaultDatasource.getInstance();
-      await vault.setSeed(widget.seed!);
-
-      final originPrivateKey = sl.get<ApiService>().getOriginKey();
-
-      await KeychainUtil().createKeyChain(
-        ref.read(SettingsProviders.settings).network,
-        widget.seed,
-        widget.name,
-        originPrivateKey,
-      );
-    } catch (e) {
-      final localizations = AppLocalizations.of(context)!;
-
-      UIUtil.showSnackbar(
-        '${localizations.sendError} ($e)',
-        context,
-        ref,
-        ArchethicTheme.text,
-        ArchethicTheme.snackBarShadow,
-      );
-    }
   }
 }
